requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pdfplumber>=0.10.2
python-dotenv>=1.0.0
openai>=1.6.0
faker>=19.6.2